pyproject.toml
[project]
name = "base-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.129.0",
    "jwt>=1.4.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.46",
]
